refactor(bot): move GitHub repo helpers into the GitHub platform module
`getGitHubRepositoryReference` and `isGitHubRepositoryLinked` are
GitHub-specific implementation details of `canHandleMessage`. Keep them
co-located with the GitHub `BotPlatform` instead of leaking into the
platform-agnostic `platform-helpers.ts`. Move their unit tests from
`platform-helpers.test.ts` into `platforms/github.test.ts` to match.
